+++ src/EncodeRom.ts
@@ -56,6 +56,65 @@ const findFileOffset = (rom: Buffer, file: Buffer) => {
   return -1;
 };
 
+const findPointerTable = (rom: Buffer) => {
+  const PLAYER_OFFSET = 0x110800;
+  const body = 0x80;
+  const head = 0xb60;
+  const rightArm = 0x26f0;
+  const eof = 0x2b40;
+  const leftArm = 0x1dd0;
+  const buster = 0x2220;
+  const feet = 0x1800;
+
+  for (let i = 0; i < rom.length - 0x20; i += 4) {
+    const a = (rom.readUInt32LE(i + 0) & 0xffffff) - PLAYER_OFFSET;
+    const b = (rom.readUInt32LE(i + 4) & 0xffffff) - PLAYER_OFFSET;
+    const c = (rom.readUInt32LE(i + 8) & 0xffffff) - PLAYER_OFFSET;
+    const d = (rom.readUInt32LE(i + 12) & 0xffffff) - PLAYER_OFFSET;
+    const e = (rom.readUInt32LE(i + 16) & 0xffffff) - PLAYER_OFFSET;
+    const f = (rom.readUInt32LE(i + 20) & 0xffffff) - PLAYER_OFFSET;
+    const g = (rom.readUInt32LE(i + 24) & 0xffffff) - PLAYER_OFFSET;
+
+    if (a !== body) {
+      continue;
+    }
+
+    if (b !== head) {
+      continue;
+    }
+
+    if (c !== rightArm) {
+      continue;
+    }
+
+    if (d !== eof) {
+      continue;
+    }
+
+    if (e !== leftArm) {
+      continue;
+    }
+
+    if (f !== buster) {
+      continue;
+    }
+
+    if (g !== feet) {
+      continue;
+    }
+
+    return i;
+  }
+
+  return -1;
+};
+
+const updatePointerTable = (rom: Buffer) => {
+  const ofs = findPointerTable(rom);
+  const dword = rom.readUint32LE(ofs);
+  rom.writeUInt32LE(dword, ofs + 4);
+};
+
 const replaceInRom = (
   sourceRom: Buffer,
   sourceFile: Buffer,
@@ -97,8 +156,8 @@ const encodeRom = () => {
   const mikuTexture = readFileSync("out/PL00T.BIN");
   const pl00t = readFileSync("bin/PL00T.BIN");
   const pl00t2 = readFileSync("bin/PL00T2.BIN");
-  replaceInRom(rom, pl00t, mikuTexture);
-  replaceInRom(rom, pl00t2, mikuTexture);
+  // replaceInRom(rom, pl00t, mikuTexture);
+  // replaceInRom(rom, pl00t2, mikuTexture);
 
   // Encode Models
   const mikuHairNorm = readFileSync("out/PL00P010.BIN");
@@ -122,9 +181,12 @@ const encodeRom = () => {
     replaceInRom(rom, file, mikuHairNorm);
   });
 
+  // Update Pointer Table
+  updatePointerTable(rom);
+
   // Write the result
   writeFileSync(romDst, rom);
 };
 
-export { encodeRom, findFileOffset };
+export { encodeRom, findFileOffset, findPointerTable };
 export default encodeRom;
